requirements.txt
Flask==2.2.2    
gunicorn==20.1.0
PyYAML==6.0
google-api-core==2.10.2   
google-auth==2.12.0   
google-cloud-bigquery==3.3.5    
google-cloud-bigquery-storage==2.16.2   
google-cloud-core==2.3.2    
google-cloud-pubsub==2.13.9   
google-cloud-storage==2.5.0  